Removed circular dependency between sr-api and job-posting.
