+++ src/job-posting.js
@@ -1,4 +1,4 @@
-import { getJobAdUrl } from './sr-api';
+import { getJobAdUrl, getJobPostingsData } from './sr-api';
 import { countryCodes } from './constants';
 
 /**
@@ -39,4 +39,20 @@ export default class JobPosting {
   get url() {
     return getJobAdUrl(this.companyId, this.uuid, this.trId);
   }
+}
+
+/**
+ * Fetches job postings from SmartRecruiters' and returns them as JobPosting
+ * objects.
+ *
+ * @param {*} companyId           The company ID (e.g. "WiproDigital")
+ * @param {*} customFieldId       Optional custom field ID to filter results on
+ * @param {*} customFieldValueId  Optional custom field value to use with the field ID
+ * @param {*} trId                Optional tracking ID
+ *
+ * @return {Promise<Array>} Promise that resolves to an array of job posting objects.
+ */
+export async function getJobPostings(companyId, customFieldId, customFieldValueId, trId) {
+  const postingsData = await getJobPostingsData(companyId, customFieldId, customFieldValueId);
+  return postingsData.map(posting => new JobPosting(posting, trId));
 }
