+++ src/job-posting.test.js
@@ -1,6 +1,10 @@
-import JobPosting from './job-posting';
-import { builditTrId as trId, countryCodes } from './constants';
+import JobPosting, { getJobPostings } from './job-posting';
 import { getJobAdUrl } from './sr-api';
+import {
+  builditTrId as trId,
+  countryCodes,
+  wiproDigitalId as companyId,
+} from './constants';
 
 const srPostingData = {
   id: '743999665860002',
@@ -121,3 +125,42 @@ describe('url getter', () => {
     ));
   });
 });
+
+
+describe('getJobPostings()', () => {
+  test('returns promise that resolves to an array', async () => {
+    const returnVal = await getJobPostings(companyId);
+    expect(Array.isArray(returnVal)).toBe(true);
+  });
+
+  test('returns only JobPosting objects', async () => {
+    const jobs = await getJobPostings(companyId);
+
+    // Try to find a job that is NOT a JobPosting instance
+    let nonJobPostingFound = false;
+    for (let i = 0; i < jobs.length; i += 1) {
+      const job = jobs[i];
+      if (!(job instanceof JobPosting)) {
+        nonJobPostingFound = true;
+        break;
+      }
+    }
+    expect(nonJobPostingFound).toBe(false);
+  });
+
+  test('passes trId to all JobPosting objects', async () => {
+    const jobs = await getJobPostings(companyId, undefined, undefined, trId);
+
+    // Try to find a job that does NOT have the trId set
+    let nonMatchingTrIdFound = false;
+    for (let i = 0; i < jobs.length; i += 1) {
+      const job = jobs[i];
+      if (job.trId !== trId) {
+        nonMatchingTrIdFound = true;
+        break;
+      }
+    }
+    expect(nonMatchingTrIdFound).toBe(false);
+  });
+});
+
